requirements.txt
openai>=1.0.0
python-dotenv
fpdf2
streamlit
requests
PyMuPDF
beautifulsoup4
python-docx
pandas